refactor: extract file related logic from ScanCodeDataProvider
